+++ tests/testthat/test-multi-cat2cat.R
@@ -9,16 +9,18 @@ occup_2008 <- occup[occup$year == 2008, ]
 occup_2010 <- occup[occup$year == 2010, ]
 occup_2012 <- occup[occup$year == 2012, ]
 
+ml_setup <- list(
+  data = occup_2010,
+  cat_var = "code",
+  method = c("knn"),
+  features = c("age", "sex", "edu", "exp", "parttime", "salary"),
+  args = list(k = 10, ntree = 50)
+)
+
 occup_back_2008_2010 <- cat2cat(
   data = list(old = occup_2008, new = occup_2010, cat_var = "code", time_var = "year"),
   mappings = list(trans = trans, direction = "backward"),
-  ml = list(
-    data = occup_2010,
-    cat_var = "code",
-    method = c("knn"),
-    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
-    args = list(k = 10, ntree = 50)
-  )
+  ml = ml_setup
 )
 
 # the counts could be any of wei_* or their combination
@@ -36,61 +38,11 @@ occup_back_2006_2008_1 <- cat2cat(
     time_var = "year"
   ),
   mappings = list(trans = trans, direction = "backward"),
-  ml = list(
-    data = occup_2010,
-    cat_var = "code",
-    method = c("knn"),
-    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
-    args = list(k = 10, ntree = 50)
-  )
+  ml = ml_setup
 )
-
-occup_2006[["index_c2c"]] <- 1:nrow(occup_2006)
-occup_2006[["g_new_c2c"]] <- occup_2006[["code"]]
-occup_back_2006_2008_2 <- cat2cat(
-  data = list(
-    old = occup_2006,
-    new = occup_back_2008_2010$old,
-    cat_var = "g_new_c2c",
-    time_var = "year"
-  ),
-  mappings = list(trans = trans, direction = "backward"),
-  ml = list(
-    data = occup_2010,
-    cat_var = "code",
-    method = c("knn"),
-    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
-    args = list(k = 10, ntree = 50)
-  )
-)
-
-occup_back_2006_2008_3 <- cat2cat(
-  data = list(
-    old = occup_2006,
-    new = occup_back_2008_2010$old,
-    cat_var = "g_new_c2c",
-    time_var = "year",
-    freqs_df = freq_df
-  ),
-  mappings = list(trans = trans, direction = "backward"),
-  ml = list(
-    data = occup_2010,
-    cat_var = "code",
-    method = c("knn"),
-    features = c("age", "sex", "edu", "exp", "parttime", "salary"),
-    args = list(k = 10, ntree = 50)
-  )
-)
-
-testthat::test_that("multi-period cat2cat freqs_df assesed internally", {
-  expect_identical(occup_back_2006_2008_1, occup_back_2006_2008_2)
-  expect_identical(occup_back_2006_2008_1, occup_back_2006_2008_3)
-})
 
 testthat::test_that("multi-period cat2cat neutral for base period", {
   expect_identical(occup_back_2008_2010$old, occup_back_2006_2008_1$new)
-  expect_identical(occup_back_2008_2010$old, occup_back_2006_2008_2$new)
-  expect_identical(occup_back_2008_2010$old, occup_back_2006_2008_3$new)
 })
 
 testthat::test_that("multi-period cat2cat probabilities", {
@@ -110,7 +62,7 @@ occup_2008_new <- occup_back_2008_2010$old # or occup_back_2006_2008$new
 occup_2010_new <- occup_back_2008_2010$new
 occup_2012_new <- dummy_c2c(occup_2012,
   cat_var = "code",
-  ml = c("wei_knn_c2c")
+  ml = c("knn")
 )
 
 final_data <- do.call(rbind, list(
@@ -136,3 +88,66 @@ counts_old <- occup %>%
 testthat::test_that("multi-period cat2cat persist the number of observations", {
   expect_identical(counts_new, counts_old)
 })
+
+data(occup)
+data(trans)
+
+occup_2006 <- occup[occup$year == 2006, ]
+occup_2008 <- occup[occup$year == 2008, ]
+occup_2010 <- occup[occup$year == 2010, ]
+occup_2012 <- occup[occup$year == 2012, ]
+
+trans2 <- rbind(trans, data.frame(old = "no_cat", new = setdiff(c(occup_2010$code, occup_2012$code), trans$new)))
+
+# 2008 -> 2010
+occup_for_2008_2010 <- cat2cat(
+  data = list(old = occup_2008, new = occup_2010, cat_var = "code", time_var = "year"),
+  mappings = list(trans = trans2, direction = "forward"),
+  ml = ml_setup
+)
+
+# optional, give more control
+# the counts could be any of wei_* or their combination
+freq_df <- occup_for_2008_2010$new[, c("g_new_c2c", "wei_freq_c2c")] %>%
+  group_by(g_new_c2c) %>%
+  summarise(counts = round(sum(wei_freq_c2c)))
+
+# 2010 -> 2012
+occup_for_2010_2012 <- cat2cat(
+  data = list(
+    old = occup_for_2008_2010$new,
+    new = occup_2012,
+    cat_var_old = "g_new_c2c",
+    cat_var_new = "code",
+    time_var = "year",
+    freqs_df = freq_df
+  ),
+  mappings = list(trans = trans2, direction = "forward"),
+  ml = ml_setup
+)
+
+# use ml argument when applied ml models
+occup_2006_new <- dummy_c2c(occup_2006, "code", ml = c("knn"))
+occup_2008_new <- occup_for_2008_2010$old
+occup_2010_new <- occup_for_2008_2010$new # or occup_for_2010_2012$old
+occup_2012_new <- occup_for_2010_2012$new
+
+final_data_for <- do.call(rbind, list(occup_2006_new, occup_2008_new, occup_2010_new, occup_2012_new))
+
+# We persist the number of observations
+counts_new <- final_data_for %>%
+  cross_c2c() %>%
+  group_by(year) %>%
+  summarise(
+    n = as.integer(round(sum(wei_freq_c2c))),
+    n2 = as.integer(round(sum(wei_cross_c2c)))
+  )
+
+counts_old <- occup %>%
+  group_by(year) %>%
+  summarise(n = n(), n2 = n(), .groups = "drop")
+
+testthat::test_that("", {
+  testthat::expect_identical(counts_new, counts_old)
+})
+